Style Update 2.3
Made changes to style, used more lines but got horizontal spaces below 99. .format() is moved to the next line pretty much everytime. This is not normal in python, but consistent with this script and works best for our needs here as we have some super long strings to deal with.