update 4.1.3
KNOWN ISSUES: Delete alias does not work. The code performs as expected, but it does not work in the gam batch operations. Also it appears to create new users sometimes when it changes the username. One user in particular shows up to be added as new user everytime the script runs. 

Other than that, runs very well. Does what we need it to do. Next version (bug fixes) should be the final version.